Make avro optional
