+++ .github/workflows/rust.yml
@@ -112,7 +112,7 @@ jobs:
           cargo test --no-default-features
           cargo run --example csv_sql
           cargo run --example parquet_sql
-          cargo run --example avro_sql
+          cargo run --example avro_sql --features=avro
         env:
           CARGO_HOME: "/github/home/.cargo"
           CARGO_TARGET_DIR: "/github/home/target"
