+++ datafusion/Cargo.toml
@@ -44,7 +44,8 @@ regex_expressions = ["regex", "lazy_static"]
 unicode_expressions = ["unicode-segmentation"]
 # Used for testing ONLY: causes all values to hash to the same value (test for collisions)
 force_hash_collisions = []
-
+# Used to enable the avro format
+avro = ["avro-rs"]
 
 [dependencies]
 ahash = "0.7"
@@ -69,7 +70,7 @@ regex = { version = "^1.4.3", optional = true }
 lazy_static = { version = "^1.4.0", optional = true }
 smallvec = { version = "1.6", features = ["union"] }
 rand = "0.8"
-avro-rs = { version = "0.13", features = ["snappy"] }
+avro-rs = { version = "0.13", features = ["snappy"], optional = true }
 num-traits = "0.2"
 
 [dev-dependencies]
